重构了UI打开逻辑，现在可以更好的进行拓展

+++ src/main/java/com/circulation/ae2wut/mixin/ae2exttable/MixinCraftingCPUStatusTWO.java
@@ -42,7 +42,7 @@ public class MixinCraftingCPUStatusTWO extends GuiCraftingCPU {
             ItemStack item = wt.getItemStack();
             if (item.getItem() instanceof ItemWirelessUniversalTerminal) {
                 if (item.getTagCompound() != null) {
-                    int mode = item.getTagCompound().getInteger("mode");
+                    byte mode = item.getTagCompound().getByte("mode");
                     var gui = ItemWirelessUniversalTerminal.getGui(mode);
                     if (gui != null) {
                         this.ae2WirelessUniversalTerminal$extendedOriginalGui = gui;
